pyproject.toml
[tool.poetry]
name = "parse_hh"
version = "0.1.0"
description = ""
authors = []

[tool.black]
line-length = 100

[tool.isort]
atomic = true
profile = "black"


[tool.poetry.dependencies]
python="3.11.4"
aiogram="^2.25.1"
httpx = "^0.24.1"
beautifulsoup4 = "^4.12.2"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
pre-commit = "^3.4.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.3.12"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-django = "^1.1.5"
flake8-quotes = "^3.3.2"
flake8-todo = "^0.7"
flake8-variables-names = "^0.0.5"
isort = "^5.12.0"
black="^23.7.0"
add-trailing-comma="^3.0.1"
